# docs/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сходится! API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/admin/danetki/settings": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "enabled",
                  "multiplayerEnabled",
                  "hostModel",
                  "promptVersion",
                  "contextMessages",
                  "maxOutputTokens",
                  "timeoutMs",
                  "retryCount",
                  "userCooldownMs",
                  "roomQuestionsPerMinute",
                  "emptyRoomTtlMinutes"
                ],
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "multiplayerEnabled": {
                    "type": "boolean"
                  },
                  "hostModel": {
                    "minLength": 1,
                    "maxLength": 120,
                    "type": "string"
                  },
                  "promptVersion": {
                    "minLength": 1,
                    "maxLength": 120,
                    "type": "string"
                  },
                  "contextMessages": {
                    "minimum": 10,
                    "maximum": 100,
                    "type": "integer"
                  },
                  "maxOutputTokens": {
                    "minimum": 100,
                    "maximum": 2000,
                    "type": "integer"
                  },
                  "timeoutMs": {
                    "minimum": 3000,
                    "maximum": 60000,
                    "type": "integer"
                  },
                  "retryCount": {
                    "minimum": 0,
                    "maximum": 3,
                    "type": "integer"
                  },
                  "userCooldownMs": {
                    "minimum": 500,
                    "maximum": 30000,
                    "type": "integer"
                  },
                  "roomQuestionsPerMinute": {
                    "minimum": 1,
                    "maximum": 120,
                    "type": "integer"
                  },
                  "emptyRoomTtlMinutes": {
                    "minimum": 1,
                    "maximum": 10080,
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/admin/danetki/sessions/{sessionId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "sessionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/danetki/sessions/{sessionId}/retry-ai": {
      "post": {
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "sessionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/danetki/sessions/{sessionId}/force-close": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "minLength": 5,
                    "maxLength": 500,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "sessionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/danetki/ai-test": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "question"
                ],
                "properties": {
                  "itemVersionId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "question": {
                    "minLength": 2,
                    "maxLength": 300,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}